+++ src/rules/no-unused-props.ts
@@ -5,41 +5,16 @@ import {
   TSESLint,
   TSESTree,
 } from "@typescript-eslint/utils";
-import { Type, TypeChecker } from "typescript";
 
 import { createRule } from "../utils/createRule";
+import { getPropertyNames } from "../utils/getPropertyNames";
+import {
+  PropsUsageTracker,
+  propsUsageTrackerFactory,
+} from "../utils/propsUsageTracker";
 import { isConstructType } from "../utils/typeCheck";
 
 type Context = TSESLint.RuleContext<"unusedProp", []>;
-
-/**
- * Tracks usage of props properties
- */
-class PropsUsageTracker {
-  private readonly propUsages = new Map<string, boolean>();
-
-  constructor(propertyNames: string[]) {
-    for (const name of propertyNames) {
-      this.propUsages.set(name, false);
-    }
-  }
-
-  markAsUsed(propertyName: string): void {
-    if (this.propUsages.has(propertyName)) {
-      this.propUsages.set(propertyName, true);
-    }
-  }
-
-  getUnusedProperties(): string[] {
-    return Array.from(this.propUsages.entries())
-      .filter(([, used]) => !used)
-      .map(([name]) => name);
-  }
-
-  hasProperties(): boolean {
-    return !!this.propUsages.size;
-  }
-}
 
 /**
  * Enforces that all properties defined in props type are used within the constructor
@@ -62,7 +37,6 @@ export const noUnusedProps = createRule({
   defaultOptions: [],
   create(context) {
     const parserServices = ESLintUtils.getParserServices(context);
-    const typeChecker = parserServices.program.getTypeChecker();
 
     return {
       ClassDeclaration(node) {
@@ -76,7 +50,7 @@ export const noUnusedProps = createRule({
         );
         if (!constructor) return;
 
-        analyzePropsUsage(constructor, context, parserServices, typeChecker);
+        analyzePropsUsage(constructor, context, parserServices);
       },
     };
   },
@@ -102,27 +76,12 @@ const reportUnusedProperties = (
 };
 
 /**
- * Creates and initializes a PropsUsageTracker
- */
-const createTracker = (
-  propsType: Type,
-  typeChecker: TypeChecker
-): PropsUsageTracker | null => {
-  const propProperties = getPropsProperties(propsType, typeChecker);
-  
-  if (!propProperties.length) return null;
-  
-  return new PropsUsageTracker(propProperties);
-};
-
-/**
  * Analyzes props usage in the constructor
  */
 const analyzePropsUsage = (
   constructor: TSESTree.MethodDefinition,
   context: Context,
-  parserServices: ParserServicesWithTypeInformation,
-  typeChecker: TypeChecker
+  parserServices: ParserServicesWithTypeInformation
 ): void => {
   const params = constructor.value.params;
 
@@ -130,131 +89,39 @@ const analyzePropsUsage = (
   if (params.length < 3) return;
 
   const propsParam = params[2];
-  
-  // Handle both identifier and object pattern (inline destructuring)
-  if (propsParam.type === AST_NODE_TYPES.Identifier) {
-    // Standard props parameter: props: MyConstructProps
-    const propsType = parserServices.getTypeAtLocation(propsParam);
-    const tracker = createTracker(propsType, typeChecker);
-    
-    if (!tracker) return;
-    
-    // Check if constructor has a body
-    if (!constructor.value.body) return;
-    
-    // Analyze constructor body for props usage
-    analyzeConstructorBody(constructor.value.body, propsParam.name, tracker);
-    
-    // Report unused properties
-    reportUnusedProperties(tracker, propsParam, context);
-  } else if (propsParam.type === AST_NODE_TYPES.ObjectPattern) {
-    // Inline destructuring: { bucketName, enableVersioning }: MyConstructProps
-    if (!propsParam.typeAnnotation?.typeAnnotation) return;
-    
-    const propsType = parserServices.getTypeAtLocation(propsParam.typeAnnotation.typeAnnotation);
-    const tracker = createTracker(propsType, typeChecker);
-    
-    if (!tracker) return;
-    
-    // Extract destructured property names
-    const destructuredProps = propsParam.properties.reduce<string[]>((acc, prop) => {
-      if (prop.type === AST_NODE_TYPES.Property && 
-          prop.key.type === AST_NODE_TYPES.Identifier) {
-        return [...acc, prop.key.name];
+
+  switch (propsParam.type) {
+    case AST_NODE_TYPES.Identifier: {
+      // NOTE: Standard props parameter (e.g. props: MyConstructProps)
+      const propsType = parserServices.getTypeAtLocation(propsParam);
+      const tracker = propsUsageTrackerFactory(propsType);
+      if (!tracker || !constructor.value.body) return;
+
+      analyzeConstructorBody(constructor.value.body, propsParam.name, tracker);
+      reportUnusedProperties(tracker, propsParam, context);
+      return;
+    }
+    case AST_NODE_TYPES.ObjectPattern: {
+      // NOTE: Inline destructuring (e.g. { bucketName, enableVersioning }: MyConstructProps)
+      if (!propsParam.typeAnnotation?.typeAnnotation) return;
+
+      const propsType = parserServices.getTypeAtLocation(
+        propsParam.typeAnnotation.typeAnnotation
+      );
+      const tracker = propsUsageTrackerFactory(propsType);
+      if (!tracker) return;
+
+      // NOTE: Mark destructured properties as used
+      for (const propName of getPropertyNames(propsParam.properties)) {
+        tracker.markAsUsed(propName);
       }
-      return acc;
-    }, []);
-    
-    // Mark destructured properties as used
-    for (const propName of destructuredProps) {
-      tracker.markAsUsed(propName);
+
+      reportUnusedProperties(tracker, propsParam, context);
+      return;
     }
-    
-    // Report unused properties
-    reportUnusedProperties(tracker, propsParam, context);
+    default:
+      return;
   }
-};
-
-/**
- * Extracts property names from props type
- */
-const getPropsProperties = (
-  propsType: Type,
-  typeChecker: TypeChecker
-): string[] => {
-  // Try to get properties from type checker first (more reliable)
-  const typeProperties = typeChecker.getPropertiesOfType(propsType);
-  if (typeProperties.length) {
-    return typeProperties.reduce<string[]>((acc, prop) => {
-      const name = prop.getName();
-      if (!isInternalProperty(name)) {
-        return [...acc, name];
-      }
-      return acc;
-    }, []);
-  }
-
-  // Fallback: get properties from symbol members
-  const symbol = propsType.getSymbol();
-  if (!symbol?.members) return [];
-
-  return Array.from(symbol.members.keys()).reduce<string[]>((acc, key) => {
-    const name = String(key);
-    if (!isInternalProperty(name)) {
-      return [...acc, name];
-    }
-    return acc;
-  }, []);
-};
-
-/**
- * Checks if a property is an internal TypeScript property
- */
-const isInternalProperty = (propertyName: string): boolean => {
-  return (
-    propertyName.startsWith("__") ||
-    propertyName === "constructor" ||
-    propertyName === "prototype"
-  );
-};
-
-/**
- * Type guard to check if a value is a TSESTree.Node
- */
-const isESTreeNode = (value: unknown): value is TSESTree.Node => {
-  return (
-    value !== null &&
-    typeof value === "object" &&
-    "type" in value &&
-    typeof (value as { type: unknown }).type === "string"
-  );
-};
-
-/**
- * Type guard to check if a value is an array of unknown values
- */
-const isUnknownArray = (value: unknown): value is unknown[] => {
-  return Array.isArray(value);
-};
-
-/**
- * Safely gets child nodes from a TSESTree.Node
- */
-const getChildNodes = (node: TSESTree.Node): TSESTree.Node[] => {
-  const skipKeys = new Set(["parent", "range", "loc"]);
-
-  return Object.entries(node).reduce<TSESTree.Node[]>((acc, [key, value]) => {
-    if (skipKeys.has(key)) return acc;
-
-    if (isESTreeNode(value)) {
-      return [...acc, value];
-    }
-    if (isUnknownArray(value)) {
-      const validNodes = value.filter(isESTreeNode);
-      return [...acc, ...validNodes];
-    }
-    return acc;
-  }, []);
 };
 
 /**
@@ -273,17 +140,17 @@ const analyzeConstructorBody = (
 
     switch (node.type) {
       case AST_NODE_TYPES.MemberExpression:
-        handleMemberExpression(node, propsParamName, tracker);
+        tracker.markAsUsedForMemberExpression(node, propsParamName);
         break;
       case AST_NODE_TYPES.VariableDeclarator:
-        handleVariableDeclarator(node, propsParamName, tracker);
+        tracker.markAsUsedForVariableDeclarator(node, propsParamName);
         break;
       case AST_NODE_TYPES.AssignmentExpression:
-        handleAssignmentExpression(node, propsParamName, tracker);
+        tracker.markAsUsedForAssignmentExpression(node, propsParamName);
         break;
     }
 
-    // Recursively visit child nodes
+    // NOTE: Recursively visit child nodes
     const children = getChildNodes(node);
     for (const child of children) {
       visitNode(child);
@@ -294,77 +161,30 @@ const analyzeConstructorBody = (
 };
 
 /**
- * Handles member expressions like props.propertyName, props?.propertyName, this.props.propertyName, or this.props?.propertyName
+ * Safely gets child nodes from a TSESTree.Node
  */
-const handleMemberExpression = (
-  node: TSESTree.MemberExpression,
-  propsParamName: string,
-  tracker: PropsUsageTracker
-): void => {
-  // Check for props.propertyName or props?.propertyName pattern
-  if (
-    node.object.type === AST_NODE_TYPES.Identifier &&
-    node.object.name === propsParamName &&
-    node.property.type === AST_NODE_TYPES.Identifier
-  ) {
-    tracker.markAsUsed(node.property.name);
-    return;
-  }
+const getChildNodes = (node: TSESTree.Node): TSESTree.Node[] => {
+  const skipKeys = new Set(["parent", "range", "loc"]);
 
-  // Check for this.props.propertyName or this.props?.propertyName pattern
-  if (node.object.type !== AST_NODE_TYPES.MemberExpression) return;
-  if (node.object.object.type !== AST_NODE_TYPES.ThisExpression) return;
-  if (node.object.property.type !== AST_NODE_TYPES.Identifier) return;
-  if (node.object.property.name !== "props") return;
-  if (node.property.type !== AST_NODE_TYPES.Identifier) return;
-
-  tracker.markAsUsed(node.property.name);
-};
-
-/**
- * Handles variable declarations with destructuring
- */
-const handleVariableDeclarator = (
-  node: TSESTree.VariableDeclarator,
-  propsParamName: string,
-  tracker: PropsUsageTracker
-): void => {
-  // Check for destructuring assignment: const { prop1, prop2 } = props
-  if (node.id.type !== AST_NODE_TYPES.ObjectPattern) return;
-  if (node.init?.type !== AST_NODE_TYPES.Identifier) return;
-  if (node.init.name !== propsParamName) return;
-
-  const propertyNames = node.id.properties.reduce<string[]>((acc, prop) => {
-    if (
-      prop.type === AST_NODE_TYPES.Property &&
-      prop.key.type === AST_NODE_TYPES.Identifier
-    ) {
-      return [...acc, prop.key.name];
+  return Object.entries(node).reduce<TSESTree.Node[]>((acc, [key, value]) => {
+    if (skipKeys.has(key)) return acc;
+    if (isESTreeNode(value)) return [...acc, value];
+    if (Array.isArray(value)) {
+      const validNodes = value.filter(isESTreeNode);
+      return [...acc, ...validNodes];
     }
     return acc;
   }, []);
-
-  for (const name of propertyNames) {
-    tracker.markAsUsed(name);
-  }
 };
 
 /**
- * Handles assignment expressions like this.props = props or this.property = props.property
+ * Type guard to check if a value is a TSESTree.Node
  */
-const handleAssignmentExpression = (
-  node: TSESTree.AssignmentExpression,
-  propsParamName: string,
-  tracker: PropsUsageTracker
-): void => {
-  // Check for this.property = props.property pattern
-  if (node.right.type !== AST_NODE_TYPES.MemberExpression) return;
-  if (node.right.object.type !== AST_NODE_TYPES.Identifier) return;
-  if (node.right.object.name !== propsParamName) return;
-  if (node.right.property.type !== AST_NODE_TYPES.Identifier) return;
-
-  tracker.markAsUsed(node.right.property.name);
-
-  // Note: this.props = props pattern doesn't mark all properties as used
-  // because we still need to check which properties are actually accessed later
+const isESTreeNode = (value: unknown): value is TSESTree.Node => {
+  return (
+    value !== null &&
+    typeof value === "object" &&
+    "type" in value &&
+    typeof (value as { type: unknown }).type === "string"
+  );
 };
