+++ src/rules/no-variable-construct-id.ts
@@ -6,7 +6,7 @@ import {
 } from "@typescript-eslint/utils";
 
 import { createRule } from "../utils/createRule";
-import { getConstructorPropertyNames } from "../utils/parseType";
+import { getConstructorPropertyNames } from "../utils/getPropertyNames";
 import { isConstructType } from "../utils/typeCheck";
 
 type Context = TSESLint.RuleContext<"invalidConstructId", []>;
