+++ src/utils/getPropertyNames.ts
@@ -1,3 +1,4 @@
+import { AST_NODE_TYPES, TSESTree } from "@typescript-eslint/utils";
 import {
   ClassDeclaration,
   ConstructorDeclaration,
@@ -7,6 +8,26 @@ import {
 } from "typescript";
 
 import { SYNTAX_KIND } from "../constants/tsInternalFlags";
+
+/**
+ * Retrieves the property names from an array of properties.
+ *
+ * @param properties An array of properties to extract names from.
+ * @returns An array of property names.
+ */
+export const getPropertyNames = (
+  properties: (TSESTree.Property | TSESTree.RestElement)[]
+): string[] => {
+  return properties.reduce<string[]>((acc, prop) => {
+    if (
+      prop.type === AST_NODE_TYPES.Property &&
+      prop.key.type === AST_NODE_TYPES.Identifier
+    ) {
+      return [...acc, prop.key.name];
+    }
+    return acc;
+  }, []);
+};
 
 /**
  * Parses type to get the property names of the class constructor.
